feat(scripts): rename
